barnes hut and convolution
the barnes-hut currently overlaps with the convolution causing double-gravity
in the range of convolution; a fix should not affect performance, so I didn't
bother since I was more interested in timings

the timings are not bad. perhaps I could even go to 4x4 leafs, but then the
barnes-hut begins to dominate.